requirements.txt
Flask==2.2.2
Flask_Login==0.6.2
mysql_connector_repackaged==0.3.1
requests==2.28.1
